# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "anvilml-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.27"
  },
  "paths": {
...
    "/v1/jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "List jobs",
        "description": "Lists all jobs, optionally filtered by status and limited by count.",
        "operationId": "list_jobs",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Optional status filter — only jobs matching this status are returned.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/JobStatus"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Optional maximum number of jobs to return.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "before",
            "in": "query",
            "description": "Cursor for future pagination — accepted but not yet passed to the\npersistence layer (forward-compatibility per `ANVILML_DESIGN.md §13.4`).\nThis field is intentionally unused by the handler — the struct field\nexists so the HTTP API accepts the query parameter without returning\na 400 error, but the handler does not forward it to `list_jobs()`.\nStored as String because serde_urlencoded cannot deserialize\nDateTime<Utc> from query strings.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error"
          }
        }
      },
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Submit a job",
        "description": "Submits a computation graph for execution. The graph is validated by the scheduler before acceptance.",
        "operationId": "submit_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Job accepted for execution",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitJobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request — malformed JSON or invalid graph"
          },
          "409": {
            "description": "Conflict — graph contains a cycle"
          },
          "500": {
            "description": "Database error"
          },
          "503": {
            "description": "Workers unavailable — no idle workers registered"
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Bulk clear terminal jobs",
        "description": "Bulk-clear terminal jobs matching the given status filter.",
        "operationId": "bulk_clear_jobs",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by job status. Must be one of: completed, failed, cancelled, all.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bulk clear completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemovedCount"
                }
              }
            }
          },
          "400": {
            "description": "Bad request — invalid status parameter"
          }
        }
      }
    },
    "/v1/jobs/{id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get a job",
        "description": "Looks up a single job by its UUID.",
        "operationId": "get_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Delete a job",
        "description": "Deletes a job by its UUID and all associated artifacts. Only terminal-status jobs may be deleted.",
        "operationId": "delete_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Job and artifacts deleted"
          },
          "404": {
            "description": "Job not found"
          },
          "409": {
            "description": "Job is not in a terminal state"
          }
        }
      }
    },
    "/v1/jobs/{id}/cancel": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Cancel a job",
        "description": "Cancels a job by its UUID. Returns 202 if accepted, 409 if already terminal, 404 if not found.",
        "operationId": "cancel_job",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Job UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Cancellation accepted"
          },
          "404": {
            "description": "Job not found"
          },
          "409": {
            "description": "Job is already in a terminal state"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Job": {
        "type": "object",
        "description": "A single generation job submitted to the AnvilML scheduler.\n\nA `Job` represents a request to run a computation graph on a worker.\nIt carries the graph definition, execution settings, lifecycle timestamps,\nand the current status. Fields that are only populated after the job\nenters a later lifecycle stage (`started_at`, `completed_at`, `worker_id`,\n`error`, `queue_position`) are `Option` types.",
        "required": [
          "id",
          "status",
          "graph",
          "settings",
          "created_at"
        ],
        "properties": {
          "completed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Timestamp when the job finished (completed or failed). None while running."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the job was first created and queued."
          },
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Error message if the job failed. Set only for `Failed` jobs."
          },
          "graph": {
            "description": "The computation graph to execute, in the format expected by workers."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Stable unique identifier for this job (UUID v4)."
          },
          "queue_position": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Position in the queue. Set when `Queued`, cleared when picked up.",
            "minimum": 0
          },
          "settings": {
            "$ref": "#/components/schemas/JobSettings",
            "description": "Optional execution settings (device preference, etc.)."
          },
          "started_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Timestamp when a worker began executing the job. None while queued."
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "description": "Current lifecycle status."
          },
          "worker_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "ID of the worker that executed this job. Set after execution begins."
          }
        }
      },
      "JobSettings": {
        "type": "object",
        "description": "Optional settings that control how a job is executed.",
        "properties": {
          "device_preference": {
            "type": [
              "string",
              "null"
            ],
            "description": "Requested device. None = auto-select by VRAM."
          }
        }
      },
      "JobStatus": {
        "type": "string",
        "description": "The lifecycle status of a generation job.\n\nJobs progress through this state machine:\n`Queued` → `Running` → (`Completed` | `Failed`) | `Cancelled` (at any point).",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed",
          "cancelled"
        ]
      },
...
      "RemovedCount": {
        "type": "object",
        "description": "HTTP response body for `DELETE /v1/jobs` on success.\n\nPer `ANVILML_DESIGN.md §13.4`: `200 { removed: u32 }`.",
        "required": [
          "removed"
        ],
        "properties": {
          "removed": {
            "type": "integer",
            "format": "int32",
            "description": "Number of jobs removed by the bulk clear operation.",
            "minimum": 0
          }
        }
      },
...
      "SubmitJobRequest": {
        "type": "object",
        "description": "HTTP request body for `POST /v1/jobs`.\n\nContains the computation graph (`serde_json::Value`) and execution\nsettings (`JobSettings`). The graph is validated by the scheduler\nbefore acceptance.",
        "required": [
          "graph",
          "settings"
        ],
        "properties": {
          "graph": {
            "description": "The computation graph to execute, in the format expected by workers."
          },
          "settings": {
            "$ref": "#/components/schemas/JobSettings",
            "description": "Optional execution settings (device preference, etc.)."
          }
        }
      },
      "SubmitJobResponse": {
        "type": "object",
        "description": "HTTP response body for `POST /v1/jobs` on success.\n\nPer `ANVILML_DESIGN.md §13.5`: `202 { job_id, queue_position }`.",
        "required": [
          "job_id",
          "queue_position"
        ],
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "description": "The UUID v4 assigned to the newly submitted job."
          },
          "queue_position": {
            "type": "integer",
            "format": "int32",
            "description": "The job's 1-based position in the submission queue at the time of\nacceptance. This is captured atomically with the enqueue operation\ninside `JobScheduler::submit()`.",
            "minimum": 0
          }
        }
      },
...
    }
  },
...
}